cleanup security agent check subcommand tests
