+++ cmd/security-agent/subcommands/compliance/command_test.go
@@ -3,23 +3,55 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build windows || !kubeapiserver
+//go:build !windows
 
 package compliance
 
 import (
 	"testing"
 
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-
 	"github.com/stretchr/testify/require"
 
 	"github.com/DataDog/datadog-agent/cmd/security-agent/command"
 	"github.com/DataDog/datadog-agent/comp/core"
+	"github.com/DataDog/datadog-agent/pkg/compliance/cli"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
-// This test suite requires the opposite build flags of the check package
-// in order to test the compliance command in environments that cannot have the check subcommand.
+// TestCheckSubcommand ultimately uses the check package, so its dependencies are different from the event subcommand
+func TestCheckSubcommand(t *testing.T) {
+	tests := []struct {
+		name     string
+		cliInput []string
+		check    func(cliParams *cli.CheckParams, params core.BundleParams)
+	}{
+		{
+			name:     "compliance check",
+			cliInput: []string{"compliance", "check"},
+			check: func(_ *cli.CheckParams, params core.BundleParams) {
+				require.Equal(t, command.LoggerName, params.LoggerName(), "logger name not matching")
+				require.Equal(t, "info", params.LogLevelFn(nil), "params.LogLevelFn not matching")
+			},
+		},
+		{
+			name:     "compliance check verbose",
+			cliInput: []string{"compliance", "check", "--verbose"},
+			check: func(_ *cli.CheckParams, params core.BundleParams) {
+				require.Equal(t, command.LoggerName, params.LoggerName(), "logger name not matching")
+				require.Equal(t, "trace", params.LogLevelFn(nil), "params.LogLevelFn not matching")
+			},
+		},
+	}
+
+	for _, test := range tests {
+		fxutil.TestOneShotSubcommand(t,
+			Commands(&command.GlobalParams{}),
+			test.cliInput,
+			cli.RunCheck,
+			test.check,
+		)
+	}
+}
 
 func TestLoadSubcommand(t *testing.T) {
 	tests := []struct {
